fix: correct return types in NativeLibrary.AlgebraicNumbers.cs
CRITICAL BUG FIX: Both algebraic number bound functions were returning
int instead of IntPtr (Z3_ast), causing incorrect behavior.

Changes:
- GetAlgebraicNumberLowerDelegate: int → IntPtr return type
- GetAlgebraicNumberUpperDelegate: int → IntPtr return type
- GetAlgebraicNumberLower: int → IntPtr return type
- GetAlgebraicNumberUpper: int → IntPtr return type

Updated XML documentation:
- Changed return descriptions from "1 on success, 0 on failure" to
  "Z3_ast handle representing rational approximation"
- Added precision details: "interval smaller than 1/10^precision"
- Clarified that result is a numeral AST of sort Real

Per Z3 C API specification (z3_api.h):
```c
Z3_ast Z3_API Z3_get_algebraic_number_lower(Z3_context c, Z3_ast a, unsigned precision);
Z3_ast Z3_API Z3_get_algebraic_number_upper(Z3_context c, Z3_ast a, unsigned precision);
```

Updated COMPARISON_AlgebraicNumbers.md to reflect the fix and change
status from "CRITICAL ISSUE" to "FIXED".

Build verified: 0 warnings, 0 errors

+++ Z3Wrap/Core/Interop/NativeLibrary.AlgebraicNumbers.cs
@@ -33,8 +33,8 @@ internal sealed partial class NativeLibrary
 
     // Delegates
 
-    private delegate int GetAlgebraicNumberLowerDelegate(IntPtr ctx, IntPtr algebraic, uint precision);
-    private delegate int GetAlgebraicNumberUpperDelegate(IntPtr ctx, IntPtr algebraic, uint precision);
+    private delegate IntPtr GetAlgebraicNumberLowerDelegate(IntPtr ctx, IntPtr algebraic, uint precision);
+    private delegate IntPtr GetAlgebraicNumberUpperDelegate(IntPtr ctx, IntPtr algebraic, uint precision);
 
     // Methods
 
@@ -44,15 +44,15 @@ internal sealed partial class NativeLibrary
     /// <param name="ctx">The Z3 context handle.</param>
     /// <param name="algebraic">The algebraic number AST.</param>
     /// <param name="precision">Decimal precision for approximation.</param>
-    /// <returns>1 on success, 0 on failure.</returns>
+    /// <returns>Z3_ast handle representing rational approximation (lower bound).</returns>
     /// <remarks>
     /// Returns rational approximation that is less than or equal to the algebraic number.
+    /// The interval isolating the algebraic number is smaller than 1/10^precision.
+    /// The result is a numeral AST of sort Real.
     /// Algebraic numbers are roots of polynomials represented symbolically in Z3.
-    /// The approximation is stored in the result parameter of the actual Z3 function.
-    /// Precision determines number of decimal digits in approximation.
     /// </remarks>
     /// <seealso href="https://z3prover.github.io/api/html/group__capi.html">Z3 C API Documentation</seealso>
-    internal int GetAlgebraicNumberLower(IntPtr ctx, IntPtr algebraic, uint precision)
+    internal IntPtr GetAlgebraicNumberLower(IntPtr ctx, IntPtr algebraic, uint precision)
     {
         var funcPtr = GetFunctionPointer("Z3_get_algebraic_number_lower");
         var func = Marshal.GetDelegateForFunctionPointer<GetAlgebraicNumberLowerDelegate>(funcPtr);
@@ -65,15 +65,15 @@ internal sealed partial class NativeLibrary
     /// <param name="ctx">The Z3 context handle.</param>
     /// <param name="algebraic">The algebraic number AST.</param>
     /// <param name="precision">Decimal precision for approximation.</param>
-    /// <returns>1 on success, 0 on failure.</returns>
+    /// <returns>Z3_ast handle representing rational approximation (upper bound).</returns>
     /// <remarks>
     /// Returns rational approximation that is greater than or equal to the algebraic number.
+    /// The interval isolating the algebraic number is smaller than 1/10^precision.
+    /// The result is a numeral AST of sort Real.
     /// Algebraic numbers are roots of polynomials represented symbolically in Z3.
-    /// The approximation is stored in the result parameter of the actual Z3 function.
-    /// Precision determines number of decimal digits in approximation.
     /// </remarks>
     /// <seealso href="https://z3prover.github.io/api/html/group__capi.html">Z3 C API Documentation</seealso>
-    internal int GetAlgebraicNumberUpper(IntPtr ctx, IntPtr algebraic, uint precision)
+    internal IntPtr GetAlgebraicNumberUpper(IntPtr ctx, IntPtr algebraic, uint precision)
     {
         var funcPtr = GetFunctionPointer("Z3_get_algebraic_number_upper");
         var func = Marshal.GetDelegateForFunctionPointer<GetAlgebraicNumberUpperDelegate>(funcPtr);
